correccion de errores generados por id en clientes y productos, correcion de servicio order generado correctamente

+++ foodAplication/src/main/java/com/globant/project/foodAplication/repository/client/IClientRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface IClientRepository extends CrudRepository<Client, Long> {
+public interface IClientRepository extends CrudRepository<Client, Integer> {
 
     Optional<Client> findByDocument(String document);
